Downloads/roofing_scraper_auto_mode/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0
playwright>=1.40.0
pandas>=2.0.0
python-dotenv>=1.0.0
retrying>=1.3.4
supabase>=2.0.0